+++ src/dom/v7400/parsed.rs
@@ -2,8 +2,9 @@
 
 use std::collections::HashMap;
 
-use crate::dom::v7400::object::scene::SceneNodeData;
-use crate::dom::v7400::object::{ObjectMeta, ObjectNodeId, SceneNodeId};
+use crate::dom::v7400::object::model::{ModelNodeData, ModelNodeId};
+use crate::dom::v7400::object::scene::{SceneNodeData, SceneNodeId};
+use crate::dom::v7400::object::{ObjectMeta, ObjectNodeId};
 
 /// Parsed node data.
 #[derive(Default, Debug, Clone, PartialEq)]
@@ -12,6 +13,8 @@ pub struct ParsedData {
     object_meta: HashMap<ObjectNodeId, ObjectMeta>,
     /// Scene nodes.
     scenes: HashMap<SceneNodeId, SceneNodeData>,
+    /// Model nodes.
+    models: HashMap<ModelNodeId, ModelNodeData>,
 }
 
 impl ParsedData {
@@ -33,5 +36,15 @@ impl ParsedData {
     /// Returns the map of scene data
     pub(crate) fn scenes_mut(&mut self) -> &mut HashMap<SceneNodeId, SceneNodeData> {
         &mut self.scenes
+    }
+
+    /// Returns the map of models data.
+    pub fn models(&self) -> &HashMap<ModelNodeId, ModelNodeData> {
+        &self.models
+    }
+
+    /// Returns the map of scene data
+    pub(crate) fn models_mut(&mut self) -> &mut HashMap<ModelNodeId, ModelNodeData> {
+        &mut self.models
     }
 }